# client/public/openapi.yaml
openapi: 3.0.3
info:
  title: WaCalls API
  version: "1.0.0"
  description: |
    API HTTP do **WaCalls** — chamadas de voz do WhatsApp pelo navegador.

    ## Autenticação
    Não há login/senha. Todas as rotas aceitam o header **`X-Client-Id`** (qualquer
    identificador estável), usado para marcar o "dono" das chamadas e filtrar eventos.

    > ⚠️ A API inicia/controla as chamadas (sinalização), mas o **áudio** exige um cliente
    > WebRTC (navegador) com captura de microfone — não é possível só com chamadas HTTP.
servers:
  - url: /
    description: Mesma origem do painel
...
security:
  - apiKeyAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key definida em WACALLS_API_KEY (vazia = sem autenticação)"
  parameters:
    ClientId:
      name: X-Client-Id
      in: header
      required: false
      schema: { type: string, example: meuapp }
      description: Identificador do cliente (dono da chamada / filtro de eventos).
    Sid:
      name: sid
      in: path
      required: true
      schema: { type: string }
      description: ID da sessão (conta WhatsApp).
...
paths:
...
  /api/sessions/{sid}/chatwoot:
    get:
      tags: [Chatwoot]
      summary: Consulta a integração Chatwoot da sessão (token omitido)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
                  chatwoot: { type: object }
    post:
      tags: [Chatwoot]
      summary: Conecta a sessão a uma inbox do Chatwoot (canal API)
      description: |
        Configura a integração. Crie no Chatwoot uma inbox do tipo **API**, pegue o
        `inbox_id`, `inbox_identifier`, o `account_id` e um `account_token` (Access Token
        do perfil/agente). Depois aponte o **webhook da inbox** para
        `POST /api/sessions/{sid}/chatwoot/webhook` para as respostas dos agentes saírem no WhatsApp.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, account_id, account_token, inbox_id]
              properties:
                url: { type: string, example: "https://chatwoot.seudominio.com" }
                account_id: { type: integer, example: 1 }
                account_token: { type: string, example: "xxxxxxxxxxxxxxxx" }
                inbox_id: { type: integer, example: 5 }
                inbox_identifier: { type: string, example: "abcd1234" }
      responses:
        "200": { description: Conectado }
        "400": { description: Campos obrigatórios faltando }
    delete:
      tags: [Chatwoot]
      summary: Desconecta a integração Chatwoot
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Desconectado }
...